ct: Refactor read-path
Remove old read path which includes 'L0_read_path' and 'interfaces'.
The new code is located in 'read_path'. The difference is that the new
code doesn't have to query the underlying partition. So the interfaces
are no longer used and can be removed.
