+++ src/v/cloud_topics/BUILD
@@ -1,14 +1,14 @@
 load("//bazel:build.bzl", "redpanda_cc_library")
 
 package(default_visibility = [
-    "//src/v/cloud_topics/L0_read_path:__pkg__",
-    "//src/v/cloud_topics/L0_read_path/tests:__pkg__",
     "//src/v/cloud_topics/batcher:__pkg__",
     "//src/v/cloud_topics/batcher/tests:__pkg__",
     "//src/v/cloud_topics/core:__pkg__",
     "//src/v/cloud_topics/core/tests:__pkg__",
     "//src/v/cloud_topics/dl_stm:__pkg__",
     "//src/v/cloud_topics/dl_stm/tests:__pkg__",
+    "//src/v/cloud_topics/read_path:__pkg__",
+    "//src/v/cloud_topics/read_path/tests:__pkg__",
     "//src/v/cloud_topics/reconciler:__pkg__",
     "//src/v/cloud_topics/reconciler/tests:__pkg__",
     "//src/v/cloud_topics/tests:__pkg__",
