+++ src/v/cloud_topics/CMakeLists.txt
@@ -10,10 +10,9 @@ v_cc_library(
     core/serializer.cc
     core/write_pipeline.cc
     core/write_request.cc
-    interfaces/cluster_partition_manager.cc
-    L0_read_path/L0_fetch_handler.cc
-    L0_read_path/placeholder_extent_reader.cc
-    L0_read_path/placeholder_extent.cc
+    read_path/fetch_request_handler.cc
+    read_path/placeholder_extent_reader.cc
+    read_path/placeholder_extent.cc
     throttler/throttler.cc
     dl_overlay.cc
     dl_snapshot.cc
@@ -28,6 +27,7 @@ v_cc_library(
   NAME cloud_topics_app
   SRCS
     app.cc
+    app_impl.cc
   DEPS
     v::cloud_topics_base
     v::cloud_topics_reconciler
