+++ src/v/cloud_topics/read_path/BUILD
@@ -2,8 +2,8 @@ load("//bazel:build.bzl", "redpanda_cc_library")
 
 package(default_visibility = [
     "//src/v/cloud_topics:__pkg__",
-    "//src/v/cloud_topics/L0_read_path:__pkg__",
-    "//src/v/cloud_topics/L0_read_path/tests:__pkg__",
+    "//src/v/cloud_topics/read_path:__pkg__",
+    "//src/v/cloud_topics/read_path/tests:__pkg__",
 ])
 
 redpanda_cc_library(
@@ -14,7 +14,7 @@ redpanda_cc_library(
     hdrs = [
         "placeholder_extent.h",
     ],
-    include_prefix = "cloud_topics/L0_read_path",
+    include_prefix = "cloud_topics/read_path",
     deps = [
         "//src/v/base",
         "//src/v/bytes",
@@ -47,7 +47,7 @@ redpanda_cc_library(
     hdrs = [
         "placeholder_extent_reader.h",
     ],
-    include_prefix = "cloud_topics/L0_read_path",
+    include_prefix = "cloud_topics/read_path",
     deps = [
         "//src/v/base",
         "//src/v/cloud_io:basic_cache_service_api",
@@ -56,7 +56,7 @@ redpanda_cc_library(
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/L0_read_path:placeholder_extent",
+        "//src/v/cloud_topics/read_path:placeholder_extent",
         "//src/v/config",
         "//src/v/model",
         "//src/v/storage",
@@ -66,14 +66,14 @@ redpanda_cc_library(
 )
 
 redpanda_cc_library(
-    name = "L0_fetch_handler",
+    name = "fetch_handler",
     srcs = [
-        "L0_fetch_handler.cc",
+        "fetch_request_handler.cc",
     ],
     hdrs = [
-        "L0_fetch_handler.h",
+        "fetch_request_handler.h",
     ],
-    include_prefix = "cloud_topics/L0_read_path",
+    include_prefix = "cloud_topics/read_path",
     deps = [
         "//src/v/base",
         "//src/v/bytes",
@@ -83,11 +83,10 @@ redpanda_cc_library(
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/L0_read_path:placeholder_extent_reader",
         "//src/v/cloud_topics/core:event_filter",
         "//src/v/cloud_topics/core:read_pipeline",
         "//src/v/cloud_topics/core:read_request",
-        "//src/v/cloud_topics/interfaces:cluster_partition_manager",
+        "//src/v/cloud_topics/read_path:placeholder_extent_reader",
         "//src/v/model",
         "//src/v/ssx:future_util",
         "//src/v/storage",
