+++ src/v/cloud_topics/read_path/README.md
