+++ src/v/cloud_topics/read_path/placeholder_extent.h
