+++ src/v/cloud_topics/read_path/placeholder_extent_reader.h
@@ -20,24 +20,6 @@
 
 namespace experimental::cloud_topics {
 
-/// Create placeholder_extent_reader instance.
-///
-/// The reader consumes another reader which returns dl_placeholder
-/// batches and materializes them on the fly.
-/// \param cfg is a log reader config
-/// \param bucket is a cloud storage bucket
-/// \param underlying is a reader that returns dl_placeholder batches
-/// \param api is a cloud_io::remote instance
-/// \param cache is a cloud storage cache instance
-/// \param rtc is a top level retry chain node
-model::record_batch_reader make_placeholder_extent_reader(
-  storage::log_reader_config cfg,
-  cloud_storage_clients::bucket_name bucket,
-  model::record_batch_reader underlying,
-  cloud_io::remote_api<ss::lowres_clock>& api,
-  cloud_io::basic_cache_service_api<ss::lowres_clock>& cache,
-  retry_chain_node& rtc);
-
 /// Consume the 'underlying' reader that returns placeholders
 /// and materialize them by downloading from the cloud storage.
 /// The method processes 'underlying' fully. The result is stored
@@ -52,7 +34,7 @@ model::record_batch_reader make_placeholder_extent_reader(
 /// \param rtc_logger is a logger that should track the progress
 ss::future<ss::circular_buffer<model::record_batch>> materialize_placeholders(
   cloud_storage_clients::bucket_name bucket,
-  model::record_batch_reader underlying,
+  ss::circular_buffer<model::record_batch> underlying,
   cloud_io::remote_api<ss::lowres_clock>& api,
   cloud_io::basic_cache_service_api<ss::lowres_clock>& cache,
   retry_chain_node& rtc,
