+++ src/v/cloud_topics/read_path/tests/BUILD
@@ -9,7 +9,7 @@ redpanda_test_cc_library(
         "mocks.h",
         "placeholder_extent_fixture.h",
     ],
-    include_prefix = "cloud_topics/L0_read_path/tests",
+    include_prefix = "cloud_topics/read_path/tests",
     visibility = ["//visibility:public"],
     deps = [
         "//src/v/base",
@@ -29,6 +29,7 @@ redpanda_test_cc_library(
         "//src/v/storage:record_batch_builder",
         "//src/v/storage:record_batch_utils",
         "//src/v/test_utils:gtest",
+        "//src/v/utils:retry_chain_node",
         "@googletest//:gtest",
         "@seastar",
     ],
@@ -42,11 +43,12 @@ redpanda_cc_gtest(
     ],
     deps = [
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/L0_read_path:placeholder_extent",
-        "//src/v/cloud_topics/L0_read_path/tests:placeholder_extent_fixture",
+        "//src/v/cloud_topics/read_path:placeholder_extent",
+        "//src/v/cloud_topics/read_path/tests:placeholder_extent_fixture",
         "//src/v/model",
         "//src/v/ssx:sformat",
         "//src/v/test_utils:gtest",
+        "//src/v/utils:retry_chain_node",
         "@googletest//:gtest",
         "@seastar",
     ],
@@ -61,33 +63,34 @@ redpanda_cc_gtest(
     deps = [
         "//src/v/base",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/L0_read_path:placeholder_extent_reader",
-        "//src/v/cloud_topics/L0_read_path/tests:placeholder_extent_fixture",
+        "//src/v/cloud_topics/read_path:placeholder_extent_reader",
+        "//src/v/cloud_topics/read_path/tests:placeholder_extent_fixture",
         "//src/v/model",
         "//src/v/ssx:sformat",
         "//src/v/test_utils:gtest",
+        "//src/v/utils:retry_chain_node",
         "@googletest//:gtest",
         "@seastar",
     ],
 )
 
 redpanda_cc_gtest(
-    name = "L0_fetch_handler_test",
+    name = "fetch_handler_test",
     timeout = "short",
     srcs = [
-        "L0_fetch_handler_test.cc",
+        "fetch_handler_test.cc",
     ],
     deps = [
         "//src/v/base",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/L0_read_path:L0_fetch_handler",
-        "//src/v/cloud_topics/L0_read_path/tests:placeholder_extent_fixture",
         "//src/v/cloud_topics/core:read_pipeline",
-        "//src/v/cloud_topics/interfaces/tests:mocks",
+        "//src/v/cloud_topics/read_path:fetch_handler",
+        "//src/v/cloud_topics/read_path/tests:placeholder_extent_fixture",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/ssx:sformat",
         "//src/v/test_utils:gtest",
+        "//src/v/utils:retry_chain_node",
         "@googletest//:gtest",
         "@seastar",
     ],
