+++ src/v/cloud_topics/read_path/tests/mocks.h
